fix(tests): resolve RAG test assertion failures
Fixed remaining test failures in RAG module:

## Type Assertion Fixes
- enrichment_integration_test.go: Changed expected types from int to int64
  for min_value/max_value assertions (SQL returns int64 by default)

## Hierarchical Indexing Fixes
- sqlite_vector_store.go: Fixed GetDocument to select and populate
  hierarchical fields (parent_id, level, summary)
- sqlite_vector_store.go: Fixed IndexDocument to insert hierarchical
  fields into database
- hierarchical_indexing_test.go: Fixed JSON deserialization for child_ids
  array ([]interface{} -> []string conversion)

## Test Results
All RAG tests now passing:
✅ TestSchemaIndexer_WithEnrichment_Integration
✅ TestHierarchicalIndexing
✅ All other RAG tests

+++ backend-go/internal/rag/enrichment_integration_test.go
@@ -185,8 +185,8 @@ func TestSchemaIndexer_WithEnrichment_Integration(t *testing.T) {
 	// Verify numeric enrichment
 	assert.Contains(t, ageDoc.Content, "range: 18 to 95")
 	assert.Contains(t, ageDoc.Content, "distinct_values: 80")
-	assert.Equal(t, 18, ageDoc.Metadata["min_value"])
-	assert.Equal(t, 95, ageDoc.Metadata["max_value"])
+	assert.Equal(t, int64(18), ageDoc.Metadata["min_value"])
+	assert.Equal(t, int64(95), ageDoc.Metadata["max_value"])
 	assert.Equal(t, 42.5, ageDoc.Metadata["avg_value"])
 
 	mockStore.AssertExpectations(t)
